requirements.txt
psycopg2==2.6.1
logbook==0.10.1
pip==1.5.4
pyyaml==3.11
